refactor(logic): move logic map to shared util directory
