+++ src/shared/util/logic.ts
@@ -1,4 +1,5 @@
 import {
+  BasicField,
   FieldResponse,
   IClientFieldSchema,
   IConditionSchema,
@@ -6,9 +7,47 @@ import {
   ILogicSchema,
   IPreventSubmitLogicSchema,
   IShowFieldsLogicSchema,
+  LogicCondition,
   LogicConditionState,
   LogicType,
 } from '../../types'
+
+const LOGIC_CONDITIONS: LogicCondition[] = [
+  [
+    BasicField.Dropdown,
+    [LogicConditionState.Equal, LogicConditionState.Either],
+  ],
+  [
+    BasicField.Number,
+    [
+      LogicConditionState.Equal,
+      LogicConditionState.Lte,
+      LogicConditionState.Gte,
+    ],
+  ],
+  [
+    BasicField.Decimal,
+    [
+      LogicConditionState.Equal,
+      LogicConditionState.Lte,
+      LogicConditionState.Gte,
+    ],
+  ],
+  [
+    BasicField.Rating,
+    [
+      LogicConditionState.Equal,
+      LogicConditionState.Lte,
+      LogicConditionState.Gte,
+    ],
+  ],
+  [BasicField.YesNo, [LogicConditionState.Equal]],
+  [BasicField.Radio, [LogicConditionState.Equal, LogicConditionState.Either]],
+]
+
+export const LOGIC_MAP = new Map<BasicField, LogicConditionState[]>(
+  LOGIC_CONDITIONS,
+)
 
 type GroupedLogic = Record<string, IConditionSchema[][]>
 export type FieldIdSet = Set<IClientFieldSchema['_id']>
